#23 월요일 followup 및 애니메이션 툴 일부 수정

+++ Jinwoo/Project/Engine/struct.h
@@ -38,7 +38,7 @@ struct tMtrlConst
 	Vec4	v4Arr[4];
 	Matrix	matArr[4];
 
-	int		bTex[TEX_PARAM::END];
+	int		bTex[(UINT)TEX_PARAM::END];
 
 	int		iPadding[2];
 };
